# openapi.yaml
openapi: 3.0.3
info:
  title: Aviation Compliance Microservice API
  description: |
    🛩️ **Enhanced Aviation Compliance Microservice**
    
    This microservice validates the regulatory compliance of aircraft models 
    against international aviation regulations (ANAC, FAA, EASA).
    
    ## Supported Aircraft Models
    - **E175 Series** - E175-E1, E175-E2 regional jets
    - **E190 Series** - E190-E1, E190-E2 regional jets  
    - **E195 Series** - E195-E1, E195-E2 regional jets
    - **Boeing 737** - Boeing 737-800 and variants
    - **Airbus A320** - A320neo and variants
    
    ## Supported Countries/Regions
    - **BRAZIL** - ANAC (Agência Nacional de Aviação Civil) regulations
    - **USA** - FAA (Federal Aviation Administration) regulations
    - **EUROPE** - EASA (European Union Aviation Safety Agency) regulations
    
    ## Features
    - ✅ Database-backed aircraft model validation
    - ✅ Authority-specific regulation retrieval
    - ✅ Enhanced E175 and E-Jets E2 support
    - ✅ Model-specific compliance validation
    - ✅ Comprehensive reporting and recommendations
    - ✅ Real-time compliance checking with optional Redis caching
    - ✅ Performance monitoring and metrics
    - ✅ Health checks and system status
    - ✅ Structured JSON logging with request correlation
    - ✅ Rate limiting protection with configurable limits
    
  version: "2.0.0"
  contact:
    name: Projeto Aviacao Team
    url: https://github.com/le-97/projetoAviacao
  license:
...
servers:
...
paths:
...
  /compliance/check/{model}/{country}:
    get:
      tags:
        - Compliance
      summary: Check aircraft compliance
      description: |
        Check compliance for an aircraft model in a specific country.
        This endpoint provides comprehensive compliance validation against
        regulatory authorities.
      operationId: check_compliance
      parameters:
        - name: model
          in: path
          required: true
          description: Aircraft model identifier
          schema:
            type: string
            enum: [E175, E175-E1, E175-E2, E190, E190-E1, E190-E2, E195, E195-E1, E195-E2, 737, A320, Phenom-300E, C-390]
          example: E175-E2
        - name: country
          in: path
          required: true
          description: Country/region for compliance check
          schema:
            type: string
            enum: [USA, BRAZIL, EUROPE]
          example: USA
      responses:
        '200':
          description: Compliance check completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceReport'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ComplianceReport:
      type: object
      required:
        - aircraft_model
        - country
        - overall_status
        - total_checks
        - compliant_checks
        - non_compliant_checks
        - critical_issues
        - checks
        - recommendations
      properties:
        aircraft_model:
          type: string
          description: The aircraft model that was checked
          example: E190
        country:
          type: string
          description: The country/region for which compliance was checked
          example: USA
        overall_status:
          type: string
          enum: [COMPLIANT, NON_COMPLIANT, PARTIAL_COMPLIANCE]
          description: Overall compliance status
        total_checks:
          type: integer
          description: Total number of checks performed
          example: 25
        compliant_checks:
          type: integer
          description: Number of compliant checks
          example: 20
        non_compliant_checks:
          type: integer
          description: Number of non-compliant checks
          example: 5
        critical_issues:
          type: integer
          description: Number of critical issues found
          example: 2
        checks:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceCheck'
          description: Individual compliance checks
        recommendations:
          type: array
          items:
            type: string
          description: Overall recommendations
        aircraft_info:
          $ref: '#/components/schemas/AircraftInfo'
        # Legacy fields
        status:
          type: string
          enum: [OK, PENDING, NON-COMPLIANT, NON_COMPLIANT]
          description: Legacy status field - use overall_status instead
          example: PENDING
        pending_requirements:
          type: array
          items:
            type: string
          description: Legacy field - use recommendations instead
          example: ["AD-2025-12: Wing inspection required"]
        compliance_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Overall compliance percentage (0-100)
          example: 85.5
        authority:
          type: string
          description: Regulatory authority name
          example: Federal Aviation Administration
        check_date:
          type: string
          format: date-time
          description: Date when compliance check was performed (ISO format)
          example: "2025-09-19T19:30:00"
    ComplianceCheck:
      type: object
      required:
        - regulation_reference
        - regulation_title
        - status
        - severity
      properties:
        regulation_reference:
          type: string
          description: Regulation reference code
          example: "FAR 25.571"
        regulation_title:
          type: string
          description: Regulation title
          example: "Damage-tolerance and fatigue evaluation of structure"
        status:
          type: string
          enum: [COMPLIANT, NON_COMPLIANT, PARTIAL_COMPLIANCE, NOT_APPLICABLE]
          description: Compliance status for this check
        severity:
          type: string
          enum: [CRITICAL, MAJOR, MINOR, INFO]
          description: Severity level of the check
        findings:
          type: array
          items:
            type: string
          description: Detailed findings
          default: []
        recommendations:
          type: array
          items:
            type: string
          description: Recommendations for compliance
          default: []
    AircraftInfo:
      type: object
      required:
        - manufacturer
        - model
      properties:
        manufacturer:
          type: string
          description: Aircraft manufacturer
          example: Embraer
        model:
          type: string
          description: Aircraft model
          example: E175-E2
        variant:
          type: string
          description: Aircraft variant
          example: E175-E2
        type_certificate:
          type: string
          description: Type certificate number
          example: A41NM
        max_seats:
          type: integer
          description: Maximum seating capacity
          example: 88
        max_weight_kg:
          type: number
          format: float
          description: Maximum weight in kg
          example: 38790.0
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
          example: UNSUPPORTED_MODEL
        message:
          type: string
          description: Human-readable error message
          example: "Aircraft model 'INVALID' is not supported"
        details:
          type: string
          description: Additional error details or suggestions
          example: "Check the API documentation for supported values"

    # New schemas for undocumented endpoints
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - {} # Allow unauthenticated access for public endpoints